return if addon webhook before changing subscription statud

+++ agile-core-classes/src/com/agilecrm/addon/AddOnUtil.java
@@ -157,7 +157,7 @@ public class AddOnUtil {
 	 */
 	public static void deleteAddOn(){
 		AddOn addOn = getAddOn();
-		if(addOn != null)
+		if(addOn != null && addOn.id != null)
 			addOn.delete();
 	}
